Add allreduce controller for PyTorch
