+++ elasticdl/python/worker/allreduce_controller.py
@@ -11,6 +11,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import os
+import socket
 import time
 import traceback
 from abc import abstractmethod
@@ -18,12 +19,8 @@ from functools import wraps
 
 from tensorflow.python.framework.errors_impl import UnknownError
 
+from elasticdl.python.common.constants import HorovodEnv
 from elasticdl.python.common.log_utils import default_logger as logger
-from elasticdl.python.worker.allreduce_trainer import (
-    DEFAULT_MAX_ALLREDUCE_RETRY_NUM,
-    DEFAULT_STEPS_TO_CHECK_RENDEZVOUS,
-    RendevousManager,
-)
 
 try:
     if os.getenv("USE_TORCH", None):
@@ -39,6 +36,60 @@ try:
 
 except ImportError:
     hvd = None
+
+
+# The default maximum number of retries for allreduce operation
+# if allreduce-based distributed training strategy is used.
+DEFAULT_MAX_ALLREDUCE_RETRY_NUM = 5
+DEFAULT_STEPS_TO_CHECK_RENDEZVOUS = 20
+
+
+class RendevousManager(object):
+    def __init__(self, master_client, master_addr):
+        self.need_broadcast = True
+        self._master_client = master_client
+        self._rendezvous_addr = master_addr
+        self._rendezvous_id = None
+
+    def init_horovod_if_needed(self):
+        self._set_horovod_env()
+        for _ in range(DEFAULT_MAX_ALLREDUCE_RETRY_NUM):
+            rank_response = self._master_client.get_comm_rank()
+            if rank_response.rank_id < 0:
+                logger.warning(
+                    "The master has not added the worker host into "
+                    "rendezvous yet. Retrying to get rank"
+                )
+                time.sleep(5)
+            else:
+                break
+
+        # If the rendezvous from master is unequal to self._rendezvous_id,
+        # the worker should rebuild the communication because the master
+        # has updated the communication group.
+        if rank_response.rendezvous_id != self._rendezvous_id:
+            logger.info("Initialize Horovod")
+            os.environ[HorovodEnv.RENDEZVOUS_PORT] = str(
+                rank_response.rendezvous_port
+            )
+            os.environ[HorovodEnv.RANK] = str(rank_response.rank_id)
+            os.environ[HorovodEnv.SIZE] = str(rank_response.world_size)
+            # Not using Horovod elastic feature in init, but need it for
+            # allreduce to call allreduce op when size=1.
+            os.environ[HorovodEnv.ELASTIC] = str(0)
+            hvd.shutdown()
+            hvd.init()
+            os.environ[HorovodEnv.ELASTIC] = str(1)
+            self._rendezvous_id = rank_response.rendezvous_id
+            self.need_broadcast = True
+
+    def _set_horovod_env(self):
+        if self._rendezvous_addr:
+            os.environ[HorovodEnv.RENDEZVOUS_ADDR] = self._rendezvous_addr
+        os.environ[HorovodEnv.CONTROLLER] = "gloo"
+        os.environ[HorovodEnv.CPU_OPERATIONS] = "gloo"
+        domain_ip = socket.gethostbyname(socket.gethostname())
+        os.environ[HorovodEnv.HOSTNAME] = domain_ip
 
 
 class AllReduceController(object):
@@ -63,9 +114,12 @@ class AllReduceController(object):
         def wrapper(*args, **kwargs):
             self._init_variables_before_first_calling(func, *args, **kwargs)
             self._init_horovod_periodically()
-            return self._try_to_call_func(func, *args, **kwargs)
+            return self.train_one_batch_with_retries(func, *args, **kwargs)
 
         return wrapper
+
+    def init_horvod_locally(self):
+        hvd.init()
 
     def _init_variables_before_first_calling(self, func, *args, **kwargs):
         if self._first_call:
@@ -86,28 +140,9 @@ class AllReduceController(object):
             self.broadcast()
             self._rendezvous_manager.need_broadcast = False
 
-    def _try_to_call_func(self, func, *args, **kwargs):
-        for _ in range(DEFAULT_MAX_ALLREDUCE_RETRY_NUM + 1):
-            try:
-                self._broadcast_if_needed()
-                result = func(*args, **kwargs)
-                self._step += 1
-                return result
-            except UnknownError as e:
-                logger.warning(
-                    "Failed to perform allreduce operation on "
-                    "the gradients. Retrying..."
-                )
-                # Those error message show that the communication
-                # to merge gradient fails and we can rebuild the
-                # communication.
-                if (
-                    "HorovodAllreduce" in e.message
-                    or "HorovodAllgather" in e.message
-                    or "HorovodBroadcast" in e.message
-                ):
-                    time.sleep(3)
-                    self._rendezvous_manager.init_horovod_if_needed()
+    @abstractmethod
+    def train_one_batch_with_retries(self, func, *args, **kwargs):
+        pass
 
     @abstractmethod
     def broadcast(self):
@@ -136,6 +171,29 @@ class TensorFlowV2AllReduceController(AllReduceController):
         broadcast_variables(self._model.variables, root_rank=0)
         broadcast_variables(self._optimizer.variables(), root_rank=0)
 
+    def train_one_batch_with_retries(self, func, *args, **kwargs):
+        for _ in range(DEFAULT_MAX_ALLREDUCE_RETRY_NUM + 1):
+            try:
+                self._broadcast_if_needed()
+                result = func(*args, **kwargs)
+                self._step += 1
+                return result
+            except UnknownError as e:
+                logger.warning(
+                    "Failed to perform allreduce operation on "
+                    "the gradients. Retrying..."
+                )
+                # Those error message show that the communication
+                # to merge gradient fails and we can rebuild the
+                # communication.
+                if (
+                    "HorovodAllreduce" in e.message
+                    or "HorovodAllgather" in e.message
+                    or "HorovodBroadcast" in e.message
+                ):
+                    time.sleep(3)
+                    self._rendezvous_manager.init_horovod_if_needed()
+
 
 class PyTorchAllReduceController(AllReduceController):
     def __init__(self, master_client, master_addr):
@@ -155,7 +213,7 @@ class PyTorchAllReduceController(AllReduceController):
         broadcast_parameters(self._model.state_dict(), root_rank=0)
         broadcast_optimizer_state(self._optimizer, root_rank=0)
 
-    def try_to_call_func(self, func, *args, **kwargs):
+    def train_one_batch_with_retries(self, func, *args, **kwargs):
         for _ in range(DEFAULT_MAX_ALLREDUCE_RETRY_NUM + 1):
             try:
                 self._broadcast_if_needed()
